% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reusable-params.R
\name{fna}
\alias{fna}
\title{Function a}
\usage{
fna(b_result1, b_result2, b_result3)
}
\arguments{
\item{b_result1}{The result from \code{\link[=fnb]{fnb()}}.}

\item{b_result2}{The result from \code{\link[=fnb]{fnb()}}.}

\item{b_result3}{The result from \href{https://google.com}{fnb()}.}

\item{b_result4}{The result from \href{https://roxygen2.r-lib.org/articles/reuse.html#child-documents}{fnb()}.}

\item{roxygen_reference}{Something referencing \href{https://roxygen2.r-lib.org/reference/load.html}{roxygen2::load_source()}.}
}
\value{
Things
}
\description{
A function to document params.
}
\details{
More details. And then I'll load details.
The result from \href{https://roxygen2.r-lib.org/articles/reuse.html#child-documents}{fnb()}.

Something referencing \href{https://roxygen2.r-lib.org/reference/load.html}{roxygen2::load_source()}.
}
